WIP: follower system port - partial implementation

+++ include/field_player_avatar.h
@@ -5,7 +5,7 @@
 
 void ClearPlayerAvatarInfo(void);
 void SetPlayerAvatarExtraStateTransition(u16, u8);
-u8 GetPlayerAvatarGenderByGraphicsId(u16);
+u16 GetPlayerAvatarGenderByGraphicsId(u16);
 bool8 TestPlayerAvatarFlags(u8);
 void PlayerGetDestCoords(s16 *, s16 *);
 u8 GetPlayerFacingDirection(void);
